refactor: move split plotting and plindex merging to core
